requirements.txt
streamlit
openai
autogen-agentchat
autogen-ext[openai]
ag2[gemini]
ag2[openai]
matplotlib 